+++ cmd/frontend/graphqlbackend/site.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
-	"os"
 
 	graphql "github.com/graph-gophers/graphql-go"
 	"github.com/graph-gophers/graphql-go/relay"
@@ -154,49 +152,11 @@ func (r *siteConfigurationResolver) EffectiveContents(ctx context.Context) (stri
 	if err := backend.CheckCurrentUserIsSiteAdmin(ctx); err != nil {
 		return "", err
 	}
-	return conf.Raw(), nil
-}
-
-func (r *siteConfigurationResolver) PendingContents(ctx context.Context) (*string, error) {
-	// 🚨 SECURITY: The site configuration contains secret tokens and credentials,
-	// so only admins may view it.
-	if err := backend.CheckCurrentUserIsSiteAdmin(ctx); err != nil {
-		return nil, err
-	}
-
-	if !conf.IsDirty() {
-		return nil, nil
-	}
-
-	rawContents, err := ioutil.ReadFile(conf.FilePath())
-	if err != nil {
-		if os.IsNotExist(err) {
-			s := "// The site configuration file does not exist."
-			return &s, nil
-		}
-		return nil, err
-	}
-
-	s := string(rawContents)
-	return &s, nil
-}
-
-// pendingOrEffectiveContents returns pendingContents if it exists, or else effectiveContents.
-func (r *siteConfigurationResolver) pendingOrEffectiveContents(ctx context.Context) (string, error) {
-	// 🚨 SECURITY: Site admin status is checked in both r.PendingContents and r.EffectiveContents,
-	// so we don't need to check it in this method.
-	pendingContents, err := r.PendingContents(ctx)
-	if err != nil {
-		return "", err
-	}
-	if pendingContents != nil {
-		return *pendingContents, nil
-	}
-	return r.EffectiveContents(ctx)
+	return conf.DefaultServerFrontendOnly.Raw(), nil
 }
 
 func (r *siteConfigurationResolver) ValidationMessages(ctx context.Context) ([]string, error) {
-	contents, err := r.pendingOrEffectiveContents(ctx)
+	contents, err := r.EffectiveContents(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -206,14 +166,11 @@ func (r *siteConfigurationResolver) ValidationMessages(ctx context.Context) ([]s
 func (r *siteConfigurationResolver) CanUpdate() bool {
 	// We assume the is-admin check has already been performed before constructing
 	// our receiver.
-	return conf.IsWritable() && processrestart.CanRestart()
+	return processrestart.CanRestart()
 }
 
 func (r *siteConfigurationResolver) Source() string {
-	s := conf.FilePath()
-	if !conf.IsWritable() {
-		s += " (read-only)"
-	}
+	s := conf.DefaultServerFrontendOnly.FilePath()
 	return s
 }
 
@@ -225,8 +182,8 @@ func (r *schemaResolver) UpdateSiteConfiguration(ctx context.Context, args *stru
 	if err := backend.CheckCurrentUserIsSiteAdmin(ctx); err != nil {
 		return false, err
 	}
-	if err := conf.Write(args.Input); err != nil {
+	if err := conf.DefaultServerFrontendOnly.Write(args.Input); err != nil {
 		return false, err
 	}
-	return conf.NeedServerRestart(), nil
+	return conf.DefaultServerFrontendOnly.NeedServerRestart(), nil
 }