+++ cmd/frontend/internal/httpapi/internal.go
@@ -158,6 +158,16 @@ func serveReposInventory(w http.ResponseWriter, r *http.Request) error {
 	return nil
 }
 
+func serveConfigurationRawJSON(w http.ResponseWriter, r *http.Request) error {
+	rawJSON := conf.DefaultServerFrontendOnly.Raw()
+	err := json.NewEncoder(w).Encode(rawJSON)
+	if err != nil {
+		return errors.Wrap(err, "Encode")
+	}
+
+	return nil
+}
+
 func serveReposList(w http.ResponseWriter, r *http.Request) error {
 	var opt db.ReposListOptions
 	err := json.NewDecoder(r.Body).Decode(&opt)