+++ pkg/txemail/template.go
@@ -1,28 +1,19 @@
 package txemail
 
 import (
-	"bytes"
 	"html"
 	htmltemplate "html/template"
-	"io"
 	"strings"
 	texttemplate "text/template"
 
 	"github.com/microcosm-cc/bluemonday"
 	"github.com/sourcegraph/sourcegraph/pkg/markdown"
-	gophermail "gopkg.in/jpoehls/gophermail.v0"
+	"github.com/sourcegraph/sourcegraph/pkg/txemail/txtypes"
 )
-
-// Templates contains the text and HTML templates for an email.
-type Templates struct {
-	Subject string // text/template subject template
-	Text    string // text/template text body template
-	HTML    string //  html/template HTML body template
-}
 
 // MustParseTemplate calls ParseTemplate and panics if an error is returned.
 // It is intended to be called in a package init func.
-func MustParseTemplate(input Templates) ParsedTemplates {
+func MustParseTemplate(input txtypes.Templates) txtypes.ParsedTemplates {
 	pt, err := ParseTemplate(input)
 	if err != nil {
 		panic("MustParseTemplate: " + err.Error())
@@ -32,7 +23,7 @@ func MustParseTemplate(input Templates) ParsedTemplates {
 
 // MustValidate panics if the templates are unparsable, otherwise it returns
 // them unmodified.
-func MustValidate(input Templates) Templates {
+func MustValidate(input txtypes.Templates) txtypes.Templates {
 	MustParseTemplate(input)
 	return input
 }
@@ -40,7 +31,7 @@ func MustValidate(input Templates) Templates {
 // ParseTemplate is a helper func for parsing the text/template and html/template
 // templates together. In the future it will also provide common template funcs
 // and a common footer.
-func ParseTemplate(input Templates) (*ParsedTemplates, error) {
+func ParseTemplate(input txtypes.Templates) (*txtypes.ParsedTemplates, error) {
 	st, err := texttemplate.New("").Funcs(textFuncMap).Parse(strings.TrimSpace(input.Subject))
 	if err != nil {
 		return nil, err
@@ -56,41 +47,7 @@ func ParseTemplate(input Templates) (*ParsedTemplates, error) {
 		return nil, err
 	}
 
-	return &ParsedTemplates{subj: st, text: tt, html: ht}, nil
-}
-
-// ParsedTemplates contains parsed text and HTML email templates.
-type ParsedTemplates struct {
-	subj *texttemplate.Template
-	text *texttemplate.Template
-	html *htmltemplate.Template
-}
-
-func (t ParsedTemplates) render(data interface{}, m *gophermail.Message) error {
-	render := func(tmpl interface {
-		Execute(io.Writer, interface{}) error
-	}) (string, error) {
-		var buf bytes.Buffer
-		if err := tmpl.Execute(&buf, data); err != nil {
-			return "", err
-		}
-		return buf.String(), nil
-	}
-
-	var err error
-	m.Subject, err = render(t.subj)
-	if err != nil {
-		return err
-	}
-	m.Body, err = render(t.text)
-	if err != nil {
-		return err
-	}
-	m.HTMLBody, err = render(t.html)
-	if err != nil {
-		return err
-	}
-	return nil
+	return &txtypes.ParsedTemplates{Subj: st, Text: tt, Html: ht}, nil
 }
 
 var (